Removed changelog redirection to master
It creates too much administrative overhead during
a release to do either of the following:
a) Commit changelogs to a different codebase (e.g. master when releasing 2.4)
b) Merge back 2.4 to 3.0 and master just in order to have changelogs showing.

In the end, the changelogs primarily belong into the branch
which release they're associated to, everything else can
be handled via mergebacks after the fact.