Update deps to PHP 7.3
I've deployed prod to PHP 7.3 yesterday, but cron jobs are breaking in dynamodb,
because there's a newer version that's compat with PHP 7.3 (and SS 3.7).
The constraints were set a bit too tightly.

Note that this also pulls in a new release of silverstripe/toolbar
which updates it's own deps to be PHP 7.3 compat:
https://github.com/silverstripe/silverstripe-globaltoolbar/releases/tag/4.2.2